Add belonging expressions
  - Move In to belongs function
  - Add Non in expression
  - Use dark evil functional magic instead inheritance blame (@dmanzanares)
